1 - POLTERGHAST CHANGES, SPIN AROUND SELF, SPREAD SHOTS, SPEEEEEEED!!! 2 - Rev+ Skeletron Prime can only have 3 Probes active at once instead of 6. 3 - Rev+ Plantera now spawns 2 types of tentacles, evenly spread ones and the normal random ones.

+++ Projectiles/Boss/PhantomBlast.cs
@@ -51,12 +51,12 @@ namespace CalamityMod.Projectiles.Boss
                 projectile.velocity.Y = projectile.velocity.Y + 0.035f;
             }
             float scaleFactor3 = 22f;
-            int num189 = (int)Player.FindClosest(projectile.Center, 1, 1);
+            int num189 = Player.FindClosest(projectile.Center, 1, 1);
             Vector2 vector20 = Main.player[num189].Center - projectile.Center;
             vector20.Normalize();
             vector20 *= scaleFactor3;
             int num190 = 80;
-            projectile.velocity = (projectile.velocity * (float)(num190 - 1) + vector20) / (float)num190;
+            projectile.velocity = (projectile.velocity * (num190 - 1) + vector20) / num190;
             if (projectile.velocity.Length() < 14f)
             {
                 projectile.velocity.Normalize();
